+++ server/routes/api.js
@@ -5,6 +5,7 @@ var basicAuth = require('express-basic-auth');
 var bcrypt = require('bcryptjs');
 var util = require('util');
 var _ = require('underscore');
+const {pickBy} = require('lodash');
 var pluginHandler = require('../plugins/pluginHandler');
 var logger = require('../log');
 var db = require('../knex/knex.js');
@@ -756,31 +757,31 @@ router.route('/capcodes')
       var color = req.body.color || 'black';
       var icon = req.body.icon || 'question';
       var ignore = req.body.ignore || 0;
-      var pluginconf = JSON.stringify(req.body.pluginconf) || "{}";
+      var pluginconf = JSON.stringify(vaccumPluginConf(req.body.pluginconf)) || "{}";
       db.from('capcodes')
         .where('id', '=', id)
         .modify(function (queryBuilder) {
           if (id == null) {
             queryBuilder.insert({
-              id: id,
-              address: address,
-              alias: alias,
-              agency: agency,
-              color: color,
-              icon: icon,
-              ignore: ignore,
-              pluginconf: pluginconf
+              id,
+              address,
+              alias,
+              agency,
+              color,
+              icon,
+              ignore,
+              pluginconf
             })
           } else {
             queryBuilder.update({
-              id: id,
-              address: address,
-              alias: alias,
-              agency: agency,
-              color: color,
-              icon: icon,
-              ignore: ignore,
-              pluginconf: pluginconf
+              id,
+              address,
+              alias,
+              agency,
+              color,
+              icon,
+              ignore,
+              pluginconf
             })
           }
         })
@@ -904,7 +905,7 @@ router.route('/capcodes/:id')
         var color = req.body.color || 'black';
         var icon = req.body.icon || 'question';
         var ignore = req.body.ignore || 0;
-        var pluginconf = JSON.stringify(req.body.pluginconf) || "{}";
+        var pluginconf = JSON.stringify(vaccumPluginConf(req.body.pluginconf)) || "{}";
         var updateAlias = req.body.updateAlias || 0;
 
         console.time('insert');
@@ -914,25 +915,25 @@ router.route('/capcodes/:id')
           .modify(function (queryBuilder) {
             if (id == null) {
               queryBuilder.insert({
-                id: id,
-                address: address,
-                alias: alias,
-                agency: agency,
-                color: color,
-                icon: icon,
-                ignore: ignore,
-                pluginconf: pluginconf
+                id,
+                address,
+                alias,
+                agency,
+                color,
+                icon,
+                ignore,
+                pluginconf
               })
             } else {
               queryBuilder.update({
-                id: id,
-                address: address,
-                alias: alias,
-                agency: agency,
-                color: color,
-                icon: icon,
-                ignore: ignore,
-                pluginconf: pluginconf
+                id,
+                address,
+                alias,
+                agency,
+                color,
+                icon,
+                ignore,
+                pluginconf
               })
             }
           })
@@ -1138,7 +1139,7 @@ router.route('/capcodeImport')
             var color = capcode.color || 'black';
             var icon = capcode.icon || 'question';
             var ignore = capcode.ignore || 0;
-            var pluginconf = JSON.stringify(capcode.pluginconf) || "{}";
+            var pluginconf = JSON.stringify(vaccumPluginConf(capcode.pluginconf)) || "{}";
             await db('capcodes')
               .returning('id')
               .where('address', '=', address)
@@ -1149,13 +1150,13 @@ router.route('/capcodeImport')
                   return db('capcodes')
                     .where('id', '=', rows.id)
                     .update({
-                      address: address,
-                      alias: alias,
-                      agency: agency,
-                      color: color,
-                      icon: icon,
-                      ignore: ignore,
-                      pluginconf: pluginconf
+                      address,
+                      alias,
+                      agency,
+                      color,
+                      icon,
+                      ignore,
+                      pluginconf
                     })
                     .then((result) => {
                       importresults.push({
@@ -1168,20 +1169,20 @@ router.route('/capcodeImport')
                       importresults.push({
                         address: address,
                         alias: alias,
-                        result: 'failed' + err
+                        result: 'failed ' + err
                       })
                     })
                 } else {
                   //Create new alias if one didn't get returned.
                   return db('capcodes').insert({
                     id: null,
-                    address: address,
-                    alias: alias,
-                    agency: agency,
-                    color: color,
-                    icon: icon,
-                    ignore: ignore,
-                    pluginconf: pluginconf
+                    address,
+                    alias,
+                    agency,
+                    color,
+                    icon,
+                    ignore,
+                    pluginconf
                   })
                     .then((result) => {
                       importresults.push({
@@ -1495,4 +1496,16 @@ function parseJSON(json) {
     // ignore errors
   }
   return parsed;
+}
+
+/**
+ * Removes all empty objects from a plugin configuration
+ * @param {Object} pconf An object containing a key for each Plugin, holding it's configuration
+ * @returns A sanitized version of the plugin configuration object holding only plugins with values set
+ */
+function vaccumPluginConf(pconf) {
+  const cleaned = pickBy(pconf, p => {
+      return Object.keys(p).length > 0
+  })
+  return cleaned;
 }